+++ system/src/Grav/Common/Page/Types.php
@@ -32,22 +32,23 @@ class Types implements \ArrayAccess, \Iterator, \Countable
     /** @var array */
     protected $items;
     /** @var array */
-    protected $systemBlueprints;
+    protected $systemBlueprints = [];
 
     /**
      * @param string $type
      * @param Blueprint|null $blueprint
+     * @return void
      */
     public function register($type, $blueprint = null)
     {
         if (!isset($this->items[$type])) {
             $this->items[$type] = [];
-        } elseif (!$blueprint) {
+        } elseif (null === $blueprint) {
             return;
         }
 
-        if (!$blueprint && $this->systemBlueprints) {
-            $blueprint = $this->systemBlueprints[$type] ?? $this->systemBlueprints['default'];
+        if (null === $blueprint) {
+            $blueprint = $this->systemBlueprints[$type] ?? $this->systemBlueprints['default'] ?? null;
         }
 
         if ($blueprint) {
@@ -56,21 +57,27 @@ class Types implements \ArrayAccess, \Iterator, \Countable
     }
 
     /**
+     * @return void
+     */
+    public function init()
+    {
+        if (null === $this->systemBlueprints) {
+            // Register all blueprints from the blueprints stream.
+            $this->systemBlueprints = $this->findBlueprints('blueprints://pages');
+            foreach ($this->systemBlueprints as $type => $blueprint) {
+                $this->register($type);
+            }
+        }
+    }
+
+    /**
      * @param string $uri
+     * @return void
      */
     public function scanBlueprints($uri)
     {
         if (!is_string($uri)) {
             throw new InvalidArgumentException('First parameter must be URI');
-        }
-
-        if (null === $this->systemBlueprints) {
-            $this->systemBlueprints = $this->findBlueprints('blueprints://pages');
-
-            // Register default by default.
-            $this->register('default');
-
-            $this->register('external');
         }
 
         foreach ($this->findBlueprints($uri) as $type => $blueprint) {
@@ -80,6 +87,7 @@ class Types implements \ArrayAccess, \Iterator, \Countable
 
     /**
      * @param string $uri
+     * @return void
      */
     public function scanTemplates($uri)
     {